Update AirGradient-DIY.ino
Added a status check after running bme.begin to make sure the BME sensor connects. I had a problem with this, and I needed to change the port it used to a differant one, and if this check was there, it would have saved me a lot of time debugging.